+++ src/payloads/codegen.rs
@@ -222,8 +222,15 @@ fn params(params: impl Iterator<Item = impl Borrow<Param>>) -> String {
             let field = &param.name;
             let ty = &param.ty;
             let flatten = match ty {
+                Type::RawTy(s) if s == "MessageId" && field == "reply_to_message_id" => {
+                    "\n            #[serde(serialize_with = \
+                     \"crate::types::serialize_reply_to_message_id\")]"
+                }
                 Type::RawTy(s)
-                    if s == "InputSticker" || s == "TargetMessage" || s == "StickerType" =>
+                    if s == "MessageId"
+                        || s == "InputSticker"
+                        || s == "TargetMessage"
+                        || s == "StickerType" =>
                 {
                     "\n            #[serde(flatten)]"
                 }