+++ src/payloads/copy_message.rs
@@ -16,7 +16,8 @@ impl_payload! {
             /// Unique identifier for the chat where the original message was sent (or channel username in the format `@channelusername`)
             pub from_chat_id: Recipient [into],
             /// Message identifier in the chat specified in _from\_chat\_id_
-            pub message_id: i32,
+            #[serde(flatten)]
+            pub message_id: MessageId,
         }
         optional {
             /// New caption for media, 0-1024 characters after entities parsing. If not specified, the original caption is kept
@@ -34,7 +35,8 @@ impl_payload! {
             /// Protects the contents of sent messages from forwarding and saving
             pub protect_content: bool,
             /// If the message is a reply, ID of the original message
-            pub reply_to_message_id: i32,
+            #[serde(serialize_with = "crate::types::serialize_reply_to_message_id")]
+            pub reply_to_message_id: MessageId,
             /// Pass _True_, if the message should be sent even if the specified replied-to message is not found
             pub allow_sending_without_reply: bool,
             /// Additional interface options. A JSON-serialized object for an [inline keyboard], [custom reply keyboard], instructions to remove reply keyboard or to force a reply from the user.