+++ src/payloads/edit_message_media.rs
@@ -2,7 +2,7 @@
 
 use serde::Serialize;
 
-use crate::types::{InlineKeyboardMarkup, InputMedia, Message, Recipient};
+use crate::types::{InlineKeyboardMarkup, InputMedia, Message, MessageId, Recipient};
 
 impl_payload! {
     /// Use this method to edit animation, audio, document, photo, or video messages. If a message is a part of a message album, then it can be edited only to a photo or a video. Otherwise, message type can be changed arbitrarily. When inline message is edited, new file can't be uploaded. Use previously uploaded file via its file_id or specify a URL. On success, the edited Message is returned.
@@ -14,7 +14,8 @@ impl_payload! {
             /// Unique identifier for the target chat or username of the target channel (in the format `@channelusername`).
             pub chat_id: Recipient [into],
             /// Identifier of the message to edit
-            pub message_id: i32,
+            #[serde(flatten)]
+            pub message_id: MessageId,
             /// A JSON-serialized object for a new media content of the message
             pub media: InputMedia,
         }